requirements.txt
google-genai>=1.0.0
sounddevice>=0.5.0
numpy>=1.26.0
pillow>=10.0.0
pynput>=1.7.6
pystray>=0.19.5
chromadb>=0.5.0
paramiko>=3.4.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
watchdog>=4.0.0
mss>=9.0.0
